feat : member_util을 공용 util 함수로 변경
